Move social collections actions to common
