+++ src/components/track/desktop/ConnectedPlaylistTile.tsx
@@ -29,6 +29,13 @@ import {
   getTracksFromCollection
 } from 'common/store/cache/collections/selectors'
 import { getUserFromCollection } from 'common/store/cache/users/selectors'
+import {
+  saveCollection,
+  unsaveCollection,
+  repostCollection,
+  undoRepostCollection,
+  shareCollection
+} from 'common/store/social/collections/actions'
 import ArtistPopover from 'components/artist/ArtistPopover'
 import { CollectionArtwork } from 'components/track/desktop/Artwork'
 import { TrackTileSize } from 'components/track/types'
@@ -46,13 +53,6 @@ import {
   UserListEntityType
 } from 'store/application/ui/userListModal/types'
 import { getUid, getBuffering, getPlaying } from 'store/player/selectors'
-import {
-  saveCollection,
-  unsaveCollection,
-  repostCollection,
-  undoRepostCollection,
-  shareCollection
-} from 'store/social/collections/actions'
 import { AppState } from 'store/types'
 import {
   albumPage,
