+++ src/components/track/mobile/ConnectedPlaylistTile.tsx
@@ -20,6 +20,13 @@ import {
   getTracksFromCollection
 } from 'common/store/cache/collections/selectors'
 import { getUserFromCollection } from 'common/store/cache/users/selectors'
+import {
+  saveCollection,
+  unsaveCollection,
+  repostCollection,
+  undoRepostCollection,
+  shareCollection
+} from 'common/store/social/collections/actions'
 import { open } from 'common/store/ui/mobile-overflow-menu/actions'
 import {
   OverflowAction,
@@ -32,13 +39,6 @@ import { RepostType } from 'containers/reposts-page/store/types'
 import { useRecord, make } from 'store/analytics/actions'
 import { getTheme } from 'store/application/ui/theme/selectors'
 import { getUid, getBuffering, getPlaying } from 'store/player/selectors'
-import {
-  saveCollection,
-  unsaveCollection,
-  repostCollection,
-  undoRepostCollection,
-  shareCollection
-} from 'store/social/collections/actions'
 import { AppState } from 'store/types'
 import {
   albumPage,
