+++ src/containers/menu/CollectionMenu.tsx
@@ -10,9 +10,9 @@ import {
 } from 'common/models/Analytics'
 import { PlayableType, ID } from 'common/models/Identifiers'
 import { getUser } from 'common/store/cache/users/selectors'
+import * as socialActions from 'common/store/social/collections/actions'
 import * as embedModalActions from 'containers/embed-modal/store/actions'
 import { open as openEditCollectionModal } from 'store/application/ui/editPlaylistModal/slice'
-import * as socialActions from 'store/social/collections/actions'
 import { AppState } from 'store/types'
 import { albumPage, playlistPage, profilePage } from 'utils/route'
