+++ src/containers/nav/desktop/NavColumn.js
@@ -19,6 +19,7 @@ import {
   createPlaylist,
   addTrackToPlaylist
 } from 'common/store/cache/collections/actions'
+import { saveCollection } from 'common/store/social/collections/actions'
 import { saveTrack } from 'common/store/social/tracks/actions'
 import CreatePlaylistModal from 'components/create-playlist/CreatePlaylistModal'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
@@ -47,7 +48,6 @@ import * as createPlaylistModalActions from 'store/application/ui/createPlaylist
 import { getIsOpen } from 'store/application/ui/createPlaylistModal/selectors'
 import { getIsDragging } from 'store/dragndrop/selectors'
 import { makeGetCurrent } from 'store/queue/selectors'
-import { saveCollection } from 'store/social/collections/actions'
 import {
   FEED_PAGE,
   TRENDING_PAGE,
