+++ src/store/social/collections/sagas.ts
@@ -15,6 +15,7 @@ import {
 } from 'common/store/cache/collections/selectors'
 import { adjustUserField } from 'common/store/cache/users/sagas'
 import { getUser } from 'common/store/cache/users/selectors'
+import * as socialActions from 'common/store/social/collections/actions'
 import { formatShareText } from 'common/utils/formatUtil'
 import { makeUid, makeKindId } from 'common/utils/uid'
 import * as notificationActions from 'containers/notification/store/actions'
@@ -26,7 +27,6 @@ import * as confirmerActions from 'store/confirmer/actions'
 import { confirmTransaction } from 'store/confirmer/sagas'
 import { removeFromPlaylistLibrary } from 'store/playlist-library/helpers'
 import { update as updatePlaylistLibrary } from 'store/playlist-library/slice'
-import * as socialActions from 'store/social/collections/actions'
 import { albumPage, playlistPage } from 'utils/route'
 import { share } from 'utils/share'
